Disable indent in #pragma asm/endasm and #asm/#endasm sections
Some compiler supports adding inline assembly code using #asm/#endasm
directives. For instance:

    #asm
        sll     a
        jrc     sub_1f      ; subtract $1f if A.x msb is set
        ret
    sub_1f:
        xor     a,#$1f
    #endasm

Another way is to use #pragma asm/endasm directives:

    #pragma asm
        XREF     _my_var
        PUSH     #LOW(my_func)
        PUSH     _my_var+0
        RETF
    App_CallEnd:
    #pragma endasm

This commit disables parsing of these section, much like when using the
keywords *INDENT-OFF* / *INDENT-ON*.

The way we detect the end of a section is limited. Since parsing is disabled, we
cannot rely on it. Instead we do a simple string search on either the word
'#endasm' or the words '#pragma' and 'endasm'.

Ideally the search should use regular expression, and the section begin/end
keywords should be user configurable.